pyproject.toml
[tool.poetry]
name = "wordcards"
version = "0.1.0"
description = ""
authors = ["rwaenrd <963852zxc123@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.3"
django-environ = "^0.11.2"
psycopg2-binary = "^2.9.10"
djangorestframework = "^3.15.2"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.1"
flake8-bugbear = "^24.10.31"
flake8-builtins = "^2.5.0"
pep8-naming = "^0.14.1"
flake8-commas = "^4.0.0"
flake8-variables-names = "^0.0.6"
flake8-import-order = "^0.18.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
